requirements.txt
Pillow==6.0
Click>=6.0
libraqm
